requirements.txt
Django==1.10.2
django-allauth==0.27.0
django-autoslug==1.9.3
django-model-utils==2.6
django-contrib-comments==1.7.3
django-taggit==0.21.2
django-mptt==0.8.6
django-braces==1.9.0
django-crispy-forms==1.6.1
django-bootstrap-pagination==1.6.2
django-markdownx==1.6.3
django-haystack==2.5.0
django-imagekit==3.3
python-slugify==1.2.1
Pillow==3.4.0
Markdown==2.6.7
Pygments==2.1.3
pytz==2016.7
Whoosh==2.7.4
jieba==0.38


